requirements.txt
numpy
python-dotenv
scikit-learn
openai
pymongo
langchain
pypdf
flask